+++ tests/CommandQuery.AzureFunctions.Tests/V1/FakeTraceWriter.cs
@@ -1,0 +1,18 @@
+﻿#if NET472
+using System.Diagnostics;
+using Microsoft.Azure.WebJobs.Host;
+
+namespace CommandQuery.AzureFunctions.Tests.V1
+{
+    public class FakeTraceWriter : TraceWriter
+    {
+        public FakeTraceWriter() : base(TraceLevel.Off)
+        {
+        }
+
+        public override void Trace(TraceEvent traceEvent)
+        {
+        }
+    }
+}
+#endif
